+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/StatesHandlerHolder.kt
@@ -6,7 +6,7 @@ import kotlin.reflect.KClass
  * Default realization of [StatesHandler]. It will incapsulate checking of [State] type in [checkHandleable] and class
  * casting in [handleState]
  */
-class StateHandlerHolder<I : State>(
+class StatesHandlerHolder<I : State>(
     private val inputKlass: KClass<I>,
     private val strict: Boolean = false,
     private val delegateTo: StatesHandler<I>
@@ -19,9 +19,20 @@ class StateHandlerHolder<I : State>(
 
     /**
      * Calls [delegateTo] method [StatesHandler.handleState] with [state] casted to [I]. Use [checkHandleable]
-     * to be sure that this [StateHandlerHolder] will be able to handle [state]
+     * to be sure that this [StatesHandlerHolder] will be able to handle [state]
      */
     override suspend fun StatesMachine.handleState(state: State): State? {
         return delegateTo.run { handleState(state as I) }
     }
 }
+
+@Deprecated("Renamed", ReplaceWith("StatesHandlerHolder"))
+typealias StateHandlerHolder<T> = StatesHandlerHolder<T>
+
+inline fun <reified T : State> StatesHandler<T>.holder(
+    strict: Boolean = true
+) = StatesHandlerHolder(
+    T::class,
+    strict,
+    this
+)