+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/StatesMachine.kt
@@ -2,11 +2,10 @@ package dev.inmo.micro_utils.fsm.common
 
 import dev.inmo.micro_utils.coroutines.*
 import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.asFlow
 
 private suspend fun <I : State> StatesMachine.launchStateHandling(
     state: State,
-    handlers: List<StateHandlerHolder<out I>>
+    handlers: List<StatesHandlerHolder<out I>>
 ): State? {
     return handlers.firstOrNull { it.checkHandleable(state) } ?.run {
         handleState(state)
@@ -35,7 +34,7 @@ interface StatesMachine : StatesHandler<State> {
          */
         operator fun invoke(
             statesManager: StatesManager,
-            handlers: List<StateHandlerHolder<*>>
+            handlers: List<StatesHandlerHolder<*>>
         ) = DefaultStatesMachine(statesManager, handlers)
     }
 }
@@ -46,7 +45,7 @@ interface StatesMachine : StatesHandler<State> {
  */
 class DefaultStatesMachine (
     private val statesManager: StatesManager,
-    private val handlers: List<StateHandlerHolder<*>>
+    private val handlers: List<StatesHandlerHolder<*>>
 ) : StatesMachine {
     /**
      * Will call [launchStateHandling] for state handling