+++ fsm/common/src/commonMain/kotlin/dev/inmo/micro_utils/fsm/common/dsl/FSMBuilder.kt
@@ -2,25 +2,27 @@ package dev.inmo.micro_utils.fsm.common.dsl
 
 import dev.inmo.micro_utils.fsm.common.*
 import dev.inmo.micro_utils.fsm.common.managers.*
-import dev.inmo.micro_utils.fsm.common.managers.InMemoryStatesManager
 import kotlin.reflect.KClass
 
 class FSMBuilder(
-    var statesManager: StatesManager = DefaultStatesManager(InMemoryDefaultStatesManagerRepo())
+    var statesManager: StatesManager = DefaultStatesManager(InMemoryDefaultStatesManagerRepo()),
+    var defaultStateHandler: StatesHandler<State>? = StatesHandler { null }
 ) {
-    private var states = mutableListOf<StateHandlerHolder<*>>()
+    private var states = mutableListOf<StatesHandlerHolder<*>>()
 
     fun <I : State> add(kClass: KClass<I>, handler: StatesHandler<I>) {
-        states.add(StateHandlerHolder(kClass, false, handler))
+        states.add(StatesHandlerHolder(kClass, false, handler))
     }
 
     fun <I : State> addStrict(kClass: KClass<I>, handler: StatesHandler<I>) {
-        states.add(StateHandlerHolder(kClass, true, handler))
+        states.add(StatesHandlerHolder(kClass, true, handler))
     }
 
     fun build() = StatesMachine(
         statesManager,
-        states.toList()
+        states.toList().let { list ->
+            defaultStateHandler ?.let { list + it.holder(false) } ?: list
+        }
     )
 }